llm_memory_system/requirements.txt
openai==1.3.0
python-dotenv==1.0.0
